pyproject.toml
[project]
name = "magic-item-shop"
version = "0.1.0"
description = "RAG-powered D&D magic item shop chatbot"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.5.1",
    "fastapi>=0.129.2",
    "google-cloud-aiplatform>=1.138.0",
    "litellm>=1.81.14",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.41.0",
]
